Add Vertex support for chat apps
