+++ agent-backend/src/chat/agents/vertex.py
@@ -1,0 +1,5 @@
+from chat.agents.base_anthropic_vertex import BaseAnthropicVertexChatAgent
+
+
+class VertexChatAgent(BaseAnthropicVertexChatAgent):
+    pass
